+++ include/mboxLib.h
@@ -61,7 +61,7 @@ typedef int MBOX_ID;
 /* -- PROTOTYPES ----------------------------------------------- */
 extern STATUS mboxCreate ( char *name, int len, MBOX_ID *pMboxId );
 extern STATUS mboxDelete ( MBOX_ID mboxId );
-extern STATUS mboxEnd ( int taskId );
+extern STATUS mboxEnd ( long taskId );
 extern STATUS mboxFind ( char *name, MBOX_ID *pMboxId );
 extern STATUS mboxInit ( char *procName );
 extern STATUS mboxIoctl ( MBOX_ID mboxId, int codeFunc, void *pArg );
